MYR-1737: refactor code
