+++ src/components/ProfileCard/ProfileCard.interfaces.ts
@@ -3,9 +3,10 @@ import {User} from '../../interfaces/user';
 export type ProfileCardProps = {
   user?: User;
   alias?: string;
+  isMobile?: boolean;
   notificationCount: number;
-  handleSignOut: () => void;
+  handleSignOut?: () => void;
   onViewProfile: () => void;
-  onSwitchAccount: () => void;
+  onSwitchAccount?: () => void;
   onShowNotificationList: () => void;
 };
