+++ src/components/ProfileCard/index.ts
@@ -1,2 +1,3 @@
 export * from './ProfileCard';
 export * from './ProfileCard.container';
+export * from './ProfileContent';
